+++ minigu/storage/Cargo.toml
@@ -4,28 +4,27 @@ name = "minigu-storage"
 version = "0.1.0"
 
 [dev-dependencies]
-csv = "1.3.1"
-minigu-catalog = { workspace = true }
 rand = "0.9.1"
-serde_json = "1.0.140"
 serial_test = "3.2.0"
 tempfile = "3.20.0"
 walkdir = "2.5.0"
 
 [dependencies]
+bitvec = "1.0.1"
 bytes = "1.10.0"
 chrono = "0.4.40"
 crc32fast = "1.4.2"
 crossbeam-epoch = "0.9"
 crossbeam-skiplist = "0.1"
+csv = "1.3.1"
 dashmap = "6.1"
+minigu-catalog = { workspace = true }
 minigu-common = { workspace = true }
 postcard = { version = "1.1.1", features = ["alloc"] }
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.140"
 thiserror = "2.0.0"
 uuid = { version = "1.7.0", features = ["v4", "serde"] }
-
-bitvec = "1.0.1"
 
 [lints]
 workspace = true
